requirements.txt
arrow==1.1.1
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-adminlte2==0.4.3
django-bootstrap3==15.0.0
django-enumfield==2.0.2
django-extensions==3.1.3
django-filter==2.4.0
django-guardian==2.4.0
django-import-export==2.5.0
django-ipware==3.0.2
django-jsonfield==1.4.1
django-phonenumber-field==5.2.0
django-redis==5.0.0
django-utils-six==2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
phonenumbers==8.12.26
psycopg2-binary==2.9.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
